fix: ui: make the footer buttons not shift up when opening modal
It involved pulling the modals html out of the "action-footer" div and
so an indentation push was necessary, that's why there are so many
changed lines.

Note: this component is still in need of a rework. I think the code
could really benefit from a split between the paying flow, the credit
flow and the (broken) receive money flow.

Closes #55