fix: ui: fix the credit modal flow (Closes #8)
chg: ui: hide the credit form upon clicking on "Done" ("terminer")
button

chg: dev: cleanup unused html tags

Will show a message explaining the next step in the process is to click
on a link and finalize transaction in Odoo, show a button letting the
user to go through this next step.

A click on the button closes the modal and opens a new window
at the address of the order that was just created.