chg: dev: [store] rename ``accounts`` in lokapi store to ``virtualAccountTree``
To prepare the addition of a variable in the store, storing the list
of all money accounts, rename this structure to a more accurate
name. This structure is a virtual construct to help display a
simplified tree of money accounts. It mixes notions like user
accounts (branches) and money accounts notions (leaves).

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>