fix: dev: refactor accounts and transaction store mutations and action to stick to good practices
Move mutation code that would query async lokapi calls to
actions. This is to make the mutations synchrone as they should be to
ensure only one mutation is ongoing at the same time.

This implied to remove the ``in place`` management of the virtual
account tree, that minimized as much as possible changes to the
structure to avoid unecessary glitches in the UI.

But at the same time, this prepares the function ``buildVirtualAccount`` to
be able to return more naturally the list of money accounts, that is planned
in a future commit.

Rename mutation to ``setXxxx`` for clarity and consistency.

Rename action that would query async lokapi calls to ``fetchXxxx``.

Removed action ``resetTRS`` that appeared to only trigger the
retrieval of transactions.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>