new: replace ``.env`` fixed config with dynamical ``config.json``
``config.json`` can be placed in the same directory of webpack backed
files. In development mode, you must place it in ``public/config.json``.
A sample file ``public/config.sample.json`` is provided.

Note that if ``mapUrl`` is not provided in ``config.json``, the navbar
will not contain any link to maps. And if the page is directly loaded
via forcing the URL, a message is shown in console.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>