fix: [ios] fix left margin of modals which made them overflow on ios
fix: [ios] add top margin to modals to avoid overlap with menu on ios

On mobile, left margin caused modals to overflow on the right of the
screen.

On mobile again, top margin for devices that have a top screen offset 
was not set correctly and would cause the modals to show underneath the
top menu.