fix: display specific pop-up error message in case we don't receive a confirmation message
Once the transaction is successfully sent, we are checking afterwards
if it is correctly registered on the financial backend. It this check
fails, we must explain the situation clearly to the user as the
transaction was probably correctly processed, but our failing check,
should probably raise the vigilance of the user.